Separate FrameBuilder into it's own file
This is the first step toward splitting up tiling.rs into logical
units and will help to reduce merge conflicts in the future.
